+++ tests/testthat/setup.R
@@ -1,6 +1,6 @@
 library(dplyr)
 library(testthat)
-library(mrgparamtab)
+library(pmparams)
 
 skip_if_no_bbi <- function(.test_name) {
   # if bbi_version() can't find bbi, it returns ""
@@ -31,15 +31,27 @@ paramKey = dplyr::tribble(
   "SIGMA11", "Proportional", "Variance", "RV", "propErr"
 )
 
-param_path <- system.file("model/nonmem/102", package = "mrgparamtab")
+#Data for testing param table (no boot)
 
-# param_est <- bbr::read_model(system.file("model/nonmem/102", package = "mrgparamtab")) %>%
-#              bbr::model_summary() %>%
-#              bbr::param_estimates()
-# write.csv(param_est, here::here("inst", "model", "nonmem", "param_est.csv"))
-param_est <- read.csv(system.file("model/nonmem/param_est.csv", package = "mrgparamtab"))
+paramPath <- system.file("model/nonmem/102", package = "pmparams")
+paramEst <- utils::read.csv(system.file("model/nonmem/param_est.csv", package = "pmparams"))
+paramModel <- bbr::read_model(system.file("model/nonmem/102", package = "pmparams"))
 
-param_model <- bbr::read_model(system.file("model/nonmem/102", package = "mrgparamtab"))
+newDF <- define_param_table(.estimates = paramEst, .key = paramKey, .ci = 95, .zscore = NULL)
 
-newDF <- defineParamTable(.estimates = param_est, .key = paramKey, .ci = 95, .zscore = NULL)
+#Data for testing boot param table
+boot_paramEstPath <- system.file("model/nonmem/boot/data/boot-106.csv", package = "pmparams")
+boot_paramEst <- utils::read.csv(system.file("model/nonmem/boot/data/boot-106.csv", package = "pmparams"))
 
+nonboot_paramEstPath <- system.file("model/nonmem/106", package = "pmparams")
+nonboot_paramEst <- utils::read.csv(system.file("model/nonmem/nonboot_param_est.csv", package = "pmparams"))
+
+newbootDF <- pmparams::define_boot_table(.boot_estimates =boot_paramEst, .nonboot_estimates = nonboot_paramEst, .key = paramKey)
+formatBootDF <- pmparams::format_boot_table(.boot_df = newbootDF)
+
+#final output
+nonbootDF <- pmparams::define_param_table(.estimates = nonboot_paramEst, .key = paramKey)
+formatnonbootDF <- nonbootDF %>% pmparams::format_param_table()
+
+
+bootParam <-  left_join(formatnonbootDF, formatBootDF, by = c("abb", "desc"))